backend/requirements.txt
# Core Django
Django==4.2.7
djangorestframework==3.14.0
django-cors-headers==4.3.0

# Authentication
djangorestframework-simplejwt==5.3.0
PyJWT==2.8.0

# Database
psycopg2-binary==2.9.9
dj-database-url==2.1.0

# WebSocket for real-time chat
channels==4.0.0
channels-redis==4.1.0
daphne==4.0.0

# Celery for async tasks
celery==5.3.4
redis==5.0.1

# Payment Gateway
payu-python==0.1.0
razorpay==1.4.1

# Astrology Calculations
swisseph==2.10.3.2
pytz==2023.3
ephem==4.1.5

# Image Processing
Pillow==10.1.0

# API Documentation
drf-yasg==1.21.7

# Video/Audio Integration
agora-token-builder==1.0.0
twilio==8.10.0

# Utilities
python-dotenv==1.0.0
requests==2.31.0

# Email
django-ses==3.5.2

# SMS
twilio==8.10.0

# File Upload
django-storages==1.14.2
boto3==1.29.7

# Monitoring & Logging
sentry-sdk==1.39.1

# Testing
pytest==7.4.3
pytest-django==4.7.0
factory-boy==3.3.0

# Production
gunicorn==21.2.0
whitenoise==6.6.0
